Moved data fetching to views
Much easier to have both views call the backend themselves than to coordinate telling the App layer to call it and then keep the current view and not break and such.

Maybe not the best practice but also probably wouldn't be that impactful in practice. At least not at a scope this small.